If the "trim" fails, we probably still want to import data anyway.
Previous example of an error:

[2021-05-31 13:52:22] 28938 filtering diff
Traceback (most recent call last):
  File "/home/renderaccount/src/regional/trim_osc.py", line 154, in <module>
    for nd in root.xpath('//way[@id={}]/nd'.format(row[0])):
  File "src/lxml/etree.pyx", line 1577, in lxml.etree._Element.xpath
  File "src/lxml/xpath.pxi", line 307, in lxml.etree.XPathElementEvaluator.__call__
  File "src/lxml/xpath.pxi", line 227, in lxml.etree._XPathEvaluatorBase._handle_result
lxml.etree.XPathEvalError: Invalid expression
[2021-05-31 13:52:38] 28938 [error] Trim_osc error

The problem is in the data from the server.  It isn't malformed XML (xmllint is happy)
and the linted output matches exactly apart from quotes.

It doesn't appear to be a modified way with 0 nodes.